Refactor SimpleFactory Pattern's code.
